+++ collective/geolocationbehavior/interfaces.py
@@ -1,5 +1,0 @@
-from zope.interface import Interface
-
-
-class IGeolocationBehaviorLayer(Interface):
-    """Browser layer for sites with this behavior installed."""