+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 1.3 (unreleased)
 ----------------
 
+- Remove browserlayer, which isn't needed any more. Includes an Upgrade step.
+  [thet]
+
 - Switch to Attribute storage instead of Annotation storage. This is easier to
   access and needs less ZODB unpickling. It's unlikely that two behaviors share
   the same geolocation schema name. A upgrade step is provided.